Rename auth.routes to auth.auth_routes and auth.lib to auth.auth_lib
Makes them easier to find quickly in VSCode. In general, I've found that I'm leaning towards fully descriptive file names instead of using directory context as part of the name, since that's how files show up in the editor.
