+++ server/auth/__init__.py
@@ -1,8 +1,8 @@
 from flask import Blueprint
 
-from .lib import *
+from .auth_helpers import *
 
 auth = Blueprint("auth", __name__, template_folder=".")
 
 # pylint: disable=wrong-import-position,cyclic-import
-from . import routes
+from . import auth_routes
